Move request-responses to `sc-network`
