+++ client/network/light/Cargo.toml
@@ -26,6 +26,7 @@ log = "0.4.16"
 prost = "0.11"
 sp-blockchain = { version = "4.0.0-dev", path = "../../../primitives/blockchain" }
 sc-client-api = { version = "4.0.0-dev", path = "../../api" }
+sc-network = { version = "0.10.0-dev", path = "../" }
 sc-network-common = { version = "0.10.0-dev", path = "../common" }
 sc-peerset = { version = "4.0.0-dev", path = "../../peerset" }
 sp-core = { version = "7.0.0", path = "../../../primitives/core" }
