+++ client/network/sync/src/engine.rs
@@ -19,13 +19,11 @@
 //! `SyncingEngine` is the actor responsible for syncing Substrate chain
 //! to tip and keep the blockchain up to date with network updates.
 
-// Local dependencies
 use crate::{
 	service::{self, chain_sync::ToServiceCommand},
 	ChainSync, ClientError, SyncingService,
 };
 
-// External dependencies
 use codec::{Decode, DecodeAll, Encode};
 use futures::{FutureExt, Stream, StreamExt};
 use futures_timer::Delay;
@@ -35,7 +33,6 @@ use prometheus_endpoint::{
 	register, Gauge, GaugeVec, MetricSource, Opts, PrometheusError, Registry, SourcedGauge, U64,
 };
 
-// Substrate dependencies
 use sc_client_api::{BlockBackend, HeaderBackend, ProofProvider};
 use sc_consensus::import_queue::ImportQueueService;
 use sc_network::config::{NetworkConfiguration, NonDefaultSetConfig};
@@ -61,7 +58,6 @@ use sp_runtime::{
 	SaturatedConversion,
 };
 
-// `std` dependencies
 use std::{
 	collections::{HashMap, HashSet},
 	num::NonZeroUsize,
