+++ client/network/sync/src/warp_request_handler.rs
@@ -22,14 +22,16 @@ use futures::{
 	stream::StreamExt,
 };
 use log::debug;
+
+use sc_network::request_responses::{
+	IncomingRequest, OutgoingResponse, ProtocolConfig as RequestResponseConfig,
+};
 use sc_network_common::{
 	config::ProtocolId,
-	request_responses::{
-		IncomingRequest, OutgoingResponse, ProtocolConfig as RequestResponseConfig,
-	},
 	sync::warp::{EncodedProof, WarpProofRequest, WarpSyncProvider},
 };
 use sp_runtime::traits::Block as BlockT;
+
 use std::{sync::Arc, time::Duration};
 
 const MAX_RESPONSE_SIZE: u64 = 16 * 1024 * 1024;
