+++ client/network/transactions/src/lib.rs
@@ -26,16 +26,13 @@
 //! - Use [`TransactionsHandlerPrototype::build`] then [`TransactionsHandler::run`] to obtain a
 //! `Future` that processes transactions.
 
-// Local dependencies
 use crate::config::*;
 
-// External dependencies
 use codec::{Decode, Encode};
 use futures::{prelude::*, stream::FuturesUnordered};
 use libp2p::{multiaddr, PeerId};
 use log::{debug, trace, warn};
 
-// Substrate dependencies
 use prometheus_endpoint::{register, Counter, PrometheusError, Registry, U64};
 use sc_network::{
 	config::{NonDefaultSetConfig, SetConfig},
@@ -51,7 +48,6 @@ use sc_network_common::{
 use sc_utils::mpsc::{tracing_unbounded, TracingUnboundedReceiver, TracingUnboundedSender};
 use sp_runtime::traits::Block as BlockT;
 
-// `std` dependencies
 use std::{
 	collections::{hash_map::Entry, HashMap},
 	iter,
